// types/api.generated.ts
export interface components {
    schemas: {
// ...
        User: {
            id: number;
            email: string;
            /** Format: date-time */
            email_verified_at: string | null;
            role: components["schemas"]["UserRole"];
            status: components["schemas"]["UserStatus"];
            /** Format: date-time */
            created_at: string | null;
            /** Format: date-time */
            updated_at: string | null;
        };
// ...
    };
// ...
}